+++ src/timeline/index.ts
@@ -1,5 +1,5 @@
 import _Timeline from './timeline';
-import _TimelineItem from './TimelineItem';
+import _TimelineItem from './timeline-item';
 import withInstall from '../utils/withInstall';
 import { TdTimelineProps } from './type';
